Update CLI usage instructions
The instructions currently document a "local" npm installation but
describe CLI invocation as though the installed package is included in
the user's PATH environmental variable. This is an invalid assumption in
most cases and will not generally work as expected.

The most common way for end users to consume this project as a binary is
to install it globally. Update the installation instructions to use the
global approach, and simplify the ussage examples accordingly.

(Local installation is possible but far less common. Users needing that
level of control can reasonably be expected to know how to re-interpret
the instructions in that case.)